fix: changes ColorPathStrct's canva_start and canva_end's types from a byte to a short, because objects can have their size as big as the entire display, which means that a data type such as byte, wouldn't be able to hold the value of 14280 decimal for a 1920x1080 display.

+++ cguieng.h
@@ -81,8 +81,8 @@ typedef struct FlagsField {
 typedef struct ColorPathStrct {
     byte face;
     byte RGB;
-    byte canva_start;
-    byte canva_end;
+    short canva_start;
+    short canva_end;
 } colorPath_t;
 typedef struct BaseObject {
     char* canva;
